+++ changes.md
@@ -8,6 +8,7 @@
 5. Fix several I/O problems; now one could read files, produced in Geant4
 6. Implement 'e2' drawing option for TH1 class, 
    use by default 'e' option when TH1 has non-empty fSumw2
+7. Reuse statistic from histogram itself, when no axis selection done 
 
 
 ## Changes in 3.4